// src/types/languageNames.ts
export enum LanguageCodes {
    FRENCH = 'fr',
    PORTUGUESE = 'pt',
    SPANISH = 'sp',
    ROMANIAN = 'ro',
    ITALIAN = 'it',
    KOREAN = 'ko',
    DANISH = 'da',
    NORWEGIAN = 'nb'
}

export enum LanguageNames {
    FRENCH = 'French',
    PORTUGUESE = 'Portuguese (Brazil)',
    SPANISH = 'Spanish',
    ROMANIAN = 'Romanian',
    ITALIAN = 'Italian',
    KOREAN = 'Korean',
    DANISH = 'Danish',
    NORWEGIAN = 'Norwegian (Bokmål)'
}